# src/main/resources/openapi/auth-api.yaml
openapi: 3.0.3
info:
  title: StaffSync Auth Service API
  version: 1.0.0
  description: Authentication and authorisation endpoints

servers:
  - url: http://localhost:8081
    description: Local development

tags:
  - name: Auth
    description: Authentication operations
  - name: Users
    description: User management operations
  - name: Roles
    description: Role and permission management

paths:
  /auth/register:
    post:
      tags: [Auth]
      operationId: register
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Email already registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /auth/login:
    post:
      tags: [Auth]
      operationId: login
      summary: Authenticate and obtain a JWT
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /auth/me:
    get:
      tags: [Auth]
      operationId: getCurrentUser
      summary: Get the currently authenticated user
      parameters:
        - name: X-User-Id
          in: header
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Current user info
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /users:
    get:
      tags: [Users]
      operationId: listUsers
      summary: List all users (ADMIN only)
      parameters:
        - name: X-User-Role
          in: header
          required: true
          schema:
            type: string
      responses:
        '200':
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /users/{id}/role:
    post:
      tags: [Users]
      operationId: assignRole
      summary: Assign a role to a user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: X-User-Role
          in: header
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssignRoleRequest'
      responses:
        '200':
          description: Role assigned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '404':
          description: User or role not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /roles:
    get:
      tags: [Roles]
      operationId: listRoles
      summary: List all roles with their permissions
      responses:
        '200':
          description: List of roles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RoleResponse'
    post:
      tags: [Roles]
      operationId: createRole
      summary: Create a new role
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRoleRequest'
      responses:
        '201':
          description: Role created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoleResponse'
        '409':
          description: Role name already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /roles/{id}/permissions:
    put:
      tags: [Roles]
      operationId: setRolePermissions
      summary: Set the full set of permissions on a role
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SetPermissionsRequest'
      responses:
        '200':
          description: Permissions updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoleResponse'
        '404':
          description: Role not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /roles/{id}:
    delete:
      tags: [Roles]
      operationId: deleteRole
      summary: Delete a role
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Role deleted
        '404':
          description: Role not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /permissions:
    get:
      tags: [Roles]
      operationId: listPermissions
      summary: List all permissions
      responses:
        '200':
          description: List of permissions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PermissionResponse'

components:
  schemas:
    RegisterRequest:
      type: object
      required: [email, password, name]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
        name:
          type: string
        roleId:
          type: string
          format: uuid

    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string

    TokenResponse:
      type: object
      properties:
        token:
          type: string
        userId:
          type: string
          format: uuid

    LoginResponse:
      type: object
      properties:
        token:
          type: string
        userId:
          type: string
          format: uuid
        role:
          type: string
        permissions:
          type: array
          items:
            type: string

    UserResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
        name:
          type: string
        role:
          $ref: '#/components/schemas/RoleResponse'
        createdAt:
          type: string
          format: date-time

    AssignRoleRequest:
      type: object
      required: [roleId]
      properties:
        roleId:
          type: string
          format: uuid

    RoleResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        permissions:
          type: array
          items:
            $ref: '#/components/schemas/PermissionResponse'

    CreateRoleRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
        permissionIds:
          type: array
          items:
            type: string
            format: uuid

    SetPermissionsRequest:
      type: object
      required: [permissionIds]
      properties:
        permissionIds:
          type: array
          items:
            type: string
            format: uuid

    PermissionResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string

    ErrorResponse:
      type: object
      properties:
        message:
          type: string
        timestamp:
          type: string
          format: date-time